resolve issue #2.
What we've done in this case will bring up a new Duo window.  If the user proceeds in either window, they will be taken to site number 2.  If they go back to the original Duo window and perform the Duo step, the IdP will display error.jsp as we won't remember the original AuthnRequest.